provided an additional condition to see if tags are to be updated on the put of a product

+++ routes/api/product-routes.js
@@ -11,7 +11,7 @@ router.get('/', async (req, res) => {
     const productData = await Product.findAll
      // JOIN with products which have a foregin key to categories
     ({
-       include: [{ model: Category, as: 'Category' },
+       include: [{ model: Category, as: 'category' },
        { model: Tag, through: ProductTag, as: 'tagged_product' }]
     });
     
@@ -21,6 +21,8 @@ router.get('/', async (req, res) => {
     }
     res.status(200).json(productData);
   } catch (err) {
+    console.log(`Inside findAll ${err}`);
+    console.log(err.message);
     res.status(500).json(err);
   }
 });
@@ -33,7 +35,7 @@ router.get('/:id', async (req, res) => {
     const productData = await Product.findByPk
     // JOIN with products which have a foregin key to categories
     (req.params.id, {
-      include: [{ model: Category, as: 'Category' },
+      include: [{ model: Category, as: 'category' },
       { model: Tag, through: ProductTag, as: 'tagged_product' }]
     });
   
@@ -44,6 +46,8 @@ router.get('/:id', async (req, res) => {
 
     res.status(200).json(productData);
   } catch (err) {
+    console.log(`Inside findAll ${err}`);
+
     res.status(500).json(err);
   }
 });
@@ -96,28 +100,33 @@ router.put('/:id', (req, res) => {
       // get list of current tag_ids
       const productTagIds = productTags.map(({ tag_id }) => tag_id);
       // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+      // check if any tags were sent
+      if(req.body.tagIds){
+        const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
+      
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+          // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ]);
+      }
+      
     })
     .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
-      // console.log(err);
+      console.log(err);
       res.status(400).json(err);
     });
 });
